requirements.txt
tabulate[widechars]
paho-mqtt
requests